feat(webview): W2 externalize Loom and ANKH
