+++ extensions/chthonic-archive/src/monolith/loomView.ts
@@ -1,6 +1,8 @@
 // @SID: EXT_LOOMVIEW_V1
 import * as vscode from 'vscode';
 import type { EntropySnapshot } from '../entropy/entropyWorkerClient';
+import { loadWebviewHtml } from '../runtime/webviewLoader';
+import { trackWebview } from '../runtime/webviewHmrWatcher';
 import type { RustificationReport } from './rustificationScore';
 
 export class LoomViewProvider implements vscode.WebviewViewProvider, vscode.Disposable {
@@ -10,6 +12,8 @@ export class LoomViewProvider implements vscode.WebviewViewProvider, vscode.Disp
     private report: RustificationReport | null = null;
     private snapshot: EntropySnapshot | null = null;
 
+    constructor(private readonly extensionUri: vscode.Uri) {}
+
     resolveWebviewView(
         webviewView: vscode.WebviewView,
         _context: vscode.WebviewViewResolveContext,
@@ -18,8 +22,13 @@ export class LoomViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         this.view = webviewView;
         webviewView.webview.options = {
             enableScripts: true,
+            localResourceRoots: [this.extensionUri],
         };
-        webviewView.webview.html = this.buildHtml();
+        webviewView.webview.html = this.buildHtml(webviewView.webview);
+        trackWebview('loom', webviewView, () => {
+            webviewView.webview.html = this.buildHtml(webviewView.webview);
+            this.postState();
+        });
 
         webviewView.webview.onDidReceiveMessage((message: unknown) => {
             if (!message || typeof message !== 'object') {
@@ -71,216 +80,7 @@ export class LoomViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         });
     }
 
-    private buildHtml(): string {
-        const nonce = createNonce();
-        return `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="Content-Security-Policy" content="default-src 'none'; style-src 'unsafe-inline'; script-src 'nonce-${nonce}';">
-    <title>The Loom</title>
-    <style>
-        :root {
-            --bg: #050505;
-            --fg: var(--vscode-editor-foreground);
-            --muted: var(--vscode-descriptionForeground);
-            --accent: #F4C430;
-            --panel: color-mix(in srgb, var(--bg) 88%, #161616);
-            --border: color-mix(in srgb, var(--accent) 32%, transparent);
-        }
-        * { box-sizing: border-box; }
-        html, body {
-            margin: 0;
-            height: 100%;
-            background: var(--bg);
-            color: var(--fg);
-            font-family: var(--vscode-font-family, Segoe UI, sans-serif);
-        }
-        body { padding: 10px; display: flex; flex-direction: column; gap: 10px; min-height: 100%; }
-        .panel {
-            border: 1px solid var(--border);
-            border-radius: 10px;
-            background: var(--panel);
-            padding: 10px;
-        }
-        .title { margin: 0; font-size: 12px; letter-spacing: 0.08em; text-transform: uppercase; color: var(--accent); }
-        .score { font-size: 28px; margin: 8px 0 0; }
-        .meta { color: var(--muted); font-size: 11px; }
-        .list { margin: 8px 0 0; padding: 0; list-style: none; max-height: 140px; overflow: auto; }
-        .list li { font-size: 11px; color: var(--muted); padding: 2px 0; }
-        .row { display: flex; gap: 8px; }
-        .row button { min-width: 0; }
-        .health-grid {
-            margin-top: 8px;
-            display: grid;
-            grid-template-columns: repeat(2, minmax(0, 1fr));
-            gap: 8px;
-        }
-        .health-cell {
-            border: 1px solid var(--border);
-            border-radius: 8px;
-            padding: 8px;
-            background: color-mix(in srgb, var(--panel) 90%, transparent);
-        }
-        .health-label {
-            color: var(--muted);
-            font-size: 10px;
-            letter-spacing: 0.05em;
-            text-transform: uppercase;
-        }
-        .health-value {
-            margin-top: 4px;
-            font-size: 13px;
-        }
-        button {
-            flex: 1;
-            border: 1px solid var(--border);
-            border-radius: 8px;
-            background: transparent;
-            color: var(--fg);
-            padding: 8px;
-            cursor: pointer;
-        }
-        button:hover {
-            border-color: var(--accent);
-            color: var(--accent);
-        }
-    </style>
-</head>
-<body>
-    <section class="panel">
-        <h1 class="title">The Loom</h1>
-        <div id="score" class="score">0%</div>
-        <div id="tier" class="meta">Gate</div>
-    </section>
-
-    <section class="panel">
-        <div class="title">Workspace Health Stream</div>
-        <div class="health-grid">
-            <div class="health-cell">
-                <div class="health-label">Tracked Files</div>
-                <div id="health-files" class="health-value">0</div>
-            </div>
-            <div class="health-cell">
-                <div class="health-label">Average Entropy</div>
-                <div id="health-entropy" class="health-value">0%</div>
-            </div>
-            <div class="health-cell">
-                <div class="health-label">Last Scan</div>
-                <div id="health-duration" class="health-value">0 ms</div>
-            </div>
-            <div class="health-cell">
-                <div class="health-label">Updated</div>
-                <div id="health-updated" class="health-value">n/a</div>
-            </div>
-        </div>
-    </section>
-
-    <section class="panel">
-        <div class="title">Markers Present</div>
-        <ul id="present" class="list"></ul>
-    </section>
-
-    <section class="panel">
-        <div class="title">Markers Missing</div>
-        <ul id="missing" class="list"></ul>
-    </section>
-
-    <div class="row">
-        <button id="refresh">Refresh Toolchain</button>
-        <button id="rescan">Rescan Health</button>
-        <button id="heal">Self-Heal</button>
-        <button id="focus">Deep Focus</button>
-        <button id="restore">Restore Order</button>
-    </div>
-
-    <script nonce="${nonce}">
-        const vscode = acquireVsCodeApi();
-        const score = document.getElementById('score');
-        const tier = document.getElementById('tier');
-        const present = document.getElementById('present');
-        const missing = document.getElementById('missing');
-        const healthFiles = document.getElementById('health-files');
-        const healthEntropy = document.getElementById('health-entropy');
-        const healthDuration = document.getElementById('health-duration');
-        const healthUpdated = document.getElementById('health-updated');
-
-        function setList(element, values) {
-            element.innerHTML = '';
-            if (!values.length) {
-                const item = document.createElement('li');
-                item.textContent = 'none';
-                element.appendChild(item);
-                return;
-            }
-            for (const value of values) {
-                const item = document.createElement('li');
-                item.textContent = value;
-                element.appendChild(item);
-            }
-        }
-
-        function formatElapsed(epochMs) {
-            if (!epochMs || Number.isNaN(epochMs)) {
-                return 'n/a';
-            }
-            const elapsed = Math.max(0, Date.now() - epochMs);
-            if (elapsed < 1000) return 'just now';
-            const sec = Math.floor(elapsed / 1000);
-            if (sec < 60) return sec + 's ago';
-            const min = Math.floor(sec / 60);
-            if (min < 60) return min + 'm ago';
-            const hr = Math.floor(min / 60);
-            if (hr < 48) return hr + 'h ago';
-            const day = Math.floor(hr / 24);
-            return day + 'd ago';
-        }
-
-        document.getElementById('refresh').addEventListener('click', () => vscode.postMessage({ type: 'refresh' }));
-        document.getElementById('rescan').addEventListener('click', () => vscode.postMessage({ type: 'rescan' }));
-        document.getElementById('heal').addEventListener('click', () => vscode.postMessage({ type: 'heal' }));
-        document.getElementById('focus').addEventListener('click', () => vscode.postMessage({ type: 'deepFocus' }));
-        document.getElementById('restore').addEventListener('click', () => vscode.postMessage({ type: 'restoreOrder' }));
-
-        window.addEventListener('message', (event) => {
-            const message = event.data;
-            if (!message || message.type !== 'state') {
-                return;
-            }
-
-            const report = message.report;
-            if (report) {
-                score.textContent = report.score + '%';
-                tier.textContent = report.tier;
-                setList(present, report.present || []);
-                setList(missing, report.missing || []);
-            }
-
-            const snapshot = message.snapshot;
-            if (snapshot) {
-                healthFiles.textContent = String(snapshot.totalFiles || 0);
-                healthEntropy.textContent = Math.round((snapshot.averageEntropy || 0) * 100) + '%';
-                healthDuration.textContent = String(snapshot.lastScanDurationMs || 0) + ' ms';
-                healthUpdated.textContent = formatElapsed(snapshot.lastScanAt || 0);
-            } else {
-                healthFiles.textContent = '—';
-                healthEntropy.textContent = '—';
-                healthDuration.textContent = '—';
-                healthUpdated.textContent = 'entropy disabled';
-            }
-        });
-    </script>
-</body>
-</html>`;
+    private buildHtml(webview: vscode.Webview): string {
+        return loadWebviewHtml(webview, this.extensionUri, 'loom');
     }
-}
-
-function createNonce(): string {
-    const alphabet = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    let nonce = '';
-    for (let i = 0; i < 24; i += 1) {
-        nonce += alphabet.charAt(Math.floor(Math.random() * alphabet.length));
-    }
-    return nonce;
 }
